Implemented 'get' function for cix program CS102
I was able to use the majority of the code from the implemented
'put' program to implement the 'get' function which retrieves a
file from the server. Seems to be working, needs to be tested
more thoroughly though. Then I can move onto the 'rm' function and
I'll be done.