Implemented 'rm' command : CS102 Asg8
I just finished implementing the 'rm' command for the CIX program,
this allows the client to tell the server to delete a file from
the remote directory. The server send a CIX_ACK if it succeeds and
a CIX_NAK if it fails. That was the last function to implement, I
just now need to perform some testing and bug squashing.