Added README, cleaned up the code, fixed warnings.
Added an official README file for the project with usage and
project description. Cleaned up the code a bit by removing old
cout statements and TODO clauses. Fixed warning about comparisons
between signed and unsigned values. Still need to perform some
more testing before final submission, but program works for now.
There is a bug where random bytes will show up sometimes for the
put and get functions, not sure why this is happening but it will
cause me to lose points, so it needs to be fixed.